feat: baseEquals, deepEquals 추가 -> shallowEquals에 baseEquals 로직 활용
