+++ packages/lib/src/equals/index.ts
@@ -1,2 +1,3 @@
+export * from "./baseEquals";
 export * from "./shallowEquals";
 export * from "./deepEquals";
